Moved Logger to its own package, adapted all affected files
