+++ logger/logger.go
@@ -1,4 +1,4 @@
-package models
+package logger
 
 import (
 	"fmt"
